+++ clx_ratio_invoice/models/__init__.py
@@ -2,5 +2,5 @@
 # Part of Odoo, CLx Media
 # See LICENSE file for full copyright & licensing details.
 
-from . import sale_advance_payment_inv
-from . import generate_invoice_date_range
+from . import co_op_sale_order_partner
+from . import sale_order